backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter>=23.5
drf-spectacular==0.29.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.0
pillow==10.2.0
pluggy==1.6.0
psycopg2-binary==2.9.9
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-django==4.11.1
python-dotenv==1.0.0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
openpyxl>=3.1.2
sendgrid==6.11.0
dj-database-url==2.1.0
